refactor: 기본 Modal내 헤더 삭제 및 폴더 이동, 스타일 분리
