+++ client/src/components/common/Modal/Modal.styles.tsx
@@ -1,0 +1,21 @@
+import styled from '@emotion/styled';
+
+export const Wrapper = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  background-color: rgba(205, 205, 205, 0.5);
+`;
+
+export const Container = styled.div`
+  width: 426px;
+  min-height: 232px;
+  margin: auto;
+  padding: 40px 80px;
+  background-color: #ffffff;
+  border: 2px solid #0068d5;
+  border-radius: 16px;
+`;
