+++ client/src/pages/_app.tsx
@@ -2,10 +2,8 @@ import React, { useState } from 'react';
 import '@/styles/global.css';
 import type { AppProps } from 'next/app';
 import { useRouter } from 'next/router';
-
-import Head from 'next/head';
-import Header from '@/components/common/Header/Header';
-import LoginModal from '@/components/common/LoginModal';
+import Header from '@/components/common/Header';
+import LoginModal from '@/components/common/Modal/LoginModal';
 
 function TILApp({ Component, pageProps }: AppProps): JSX.Element {
   const router = useRouter();
